+++ src/Products/PloneMeeting/tests/testMeetingItem.py
@@ -8815,17 +8815,22 @@ class testMeetingItem(PloneMeetingTestCase):
         self.assertEqual(item.Title(), "My title héhé")
         self.assertEqual(item.Title(withMeetingDate=True), "My title héhé")
         self.assertEqual(item.Title(withItemNumber=True), "My title héhé")
-        self.assertEqual(item.Title(withMeetingDate=True, withItemNumber=True),
-                         "My title héhé")
+        self.assertEqual(item.Title(withItemReference=True), "My title héhé")
+        self.assertEqual(
+            item.Title(withMeetingDate=True, withItemNumber=True, withItemReference=True),
+            "My title héhé")
         self.changeUser('pmManager')
         self.create('Meeting', date=datetime(2024, 3, 27, 15, 30))
         self.presentItem(item)
+        item.update_item_reference()
         self.assertEqual(item.Title(), "My title héhé")
         self.assertEqual(item.Title(withMeetingDate=True),
                          "My title héhé (27 march 2024 (15:30))")
         self.assertEqual(item.Title(withItemNumber=True), "3. My title héhé")
-        self.assertEqual(item.Title(withMeetingDate=True, withItemNumber=True),
-                         "3. My title héhé (27 march 2024 (15:30))")
+        self.assertEqual(item.Title(withItemReference=True), "[Ref. 20240327/3] My title héhé")
+        self.assertEqual(
+            item.Title(withMeetingDate=True, withItemNumber=True, withItemReference=True),
+            "3. [Ref. 20240327/3] My title héhé (27 march 2024 (15:30))")
 
 
 def test_suite():